+++ CHANGES.rst
@@ -1,4 +1,7 @@
 Changelog
 
+- Get browser id manager directly to avoid breaking when using
+  BeakerSessionDataManager.
+
 - Trigger a CheckoutComplete event after checkout is complete
   and order has been recorded.